Testing to block try.js for Legacy ver.
This may not safe because allowing those will be sure the ads and cookie
will be run and deploy due to they have a line of code that can re-execute
ILS, the lazyload will not run if some function was fail to run.

The best of hope is let FF legacy handle the ##script:contains() to remove
the code that (not needed to) use for inject ads, then hope for magic to be
happen.