:door: To close #21
This is also not an unexpected step of this site owner, adding the code that pretending to resolve the problem, but I've exposed the inner code is still rely to ads (their main income) in order to use their site normally.
So I will patch this hole, no matter what!