Update Normal ASTLW
To response for some report of Adfilt repo (Legitimate URLs Shortener) that blocking `fb_dtsg^` parameters will prevent uploading photos on Facebook, however, those parameters still appear or show in certain actions (e.g. navigating to any `/nt/` pages) that seems to useless for consumers, even though it's related to CSRF and for securities, but it shouldn't show or expose with `GET` requests on address bar at all!

Also included in this commit
- Add support for Chocolatey for strong 3p denied, then put Disqus and it's related to `denyallow`